+++ packages/spacecat-shared-data-access/src/models/sentiment-topic/sentiment-topic.schema.js
@@ -55,11 +55,25 @@ const schema = new SchemaBuilder(SentimentTopic, SentimentTopicCollection)
     type: 'string',
     required: false,
   })
-  .addAttribute('subPrompts', {
+  .addAttribute('urls', {
     type: 'list',
-    items: { type: 'string' },
-    required: true,
+    required: false,
     default: [],
+    postgrestIgnore: true,
+    items: {
+      type: 'map',
+      properties: {
+        url: { type: 'string', required: true },
+        timesCited: { type: 'number', required: true },
+        category: { type: 'string', required: false },
+        subPrompts: {
+          type: 'list',
+          items: { type: 'string' },
+          required: false,
+          default: [],
+        },
+      },
+    },
   })
   .addAttribute('enabled', {
     type: 'boolean',