+++ packages/spacecat-shared-http-utils/src/index.js
@@ -162,6 +162,7 @@ export function internalServerError(message = 'internal server error', headers =
 }
 
 export { authWrapper } from './auth/auth-wrapper.js';
+export { s2sAuthWrapper } from './auth/s2s-wrapper.js';
 export { enrichPathInfo } from './enrich-path-info-wrapper.js';
 export { hashWithSHA256 } from './auth/generate-hash.js';