+++ packages/spacecat-shared-vault-secrets/CHANGELOG.md
@@ -1,3 +1,27 @@
+## [@adobe/spacecat-shared-vault-secrets-v1.2.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-vault-secrets-v1.1.1...@adobe/spacecat-shared-vault-secrets-v1.2.0) (2026-03-10)
+
+### Features
+
+* enable HTTP connection reuse for PostgrestClient and VaultClient ([#1423](https://github.com/adobe/spacecat-shared/issues/1423)) ([ff92207](https://github.com/adobe/spacecat-shared/commit/ff922075a7e4d035c663d33b8b16ebe1c4dce2c6))
+
+## [@adobe/spacecat-shared-vault-secrets-v1.1.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-vault-secrets-v1.1.0...@adobe/spacecat-shared-vault-secrets-v1.1.1) (2026-03-07)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1410](https://github.com/adobe/spacecat-shared/issues/1410)) ([e5b5634](https://github.com/adobe/spacecat-shared/commit/e5b5634f788a8f440ef49218c0b2fa1ccb42fb8a))
+
+## [@adobe/spacecat-shared-vault-secrets-v1.1.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-vault-secrets-v1.0.1...@adobe/spacecat-shared-vault-secrets-v1.1.0) (2026-03-04)
+
+### Features
+
+* add dev alias secret overrides to spacecat-shared-vault-secrets ([#1400](https://github.com/adobe/spacecat-shared/issues/1400)) ([93b7c67](https://github.com/adobe/spacecat-shared/commit/93b7c670d301d8ac7eae12e2e903e50541fd9f44))
+
+## [@adobe/spacecat-shared-vault-secrets-v1.0.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-vault-secrets-v1.0.0...@adobe/spacecat-shared-vault-secrets-v1.0.1) (2026-03-02)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1223](https://github.com/adobe/spacecat-shared/issues/1223)) ([7ee8461](https://github.com/adobe/spacecat-shared/commit/7ee8461c99223d07a2f47bd6838b6942fcb30f28))
+
 ## @adobe/spacecat-shared-vault-secrets-v1.0.0 (2026-02-19)
 
 ### Features